src/services/search/requirements.txt
Flask==2.3.2
google-cloud-bigquery==3.13.0
pandas==2.1.3
